requirements.txt
Keras==2.3.1
yfinance==0.1.63
matplotlib==3.3.4
numpy==1.21.4
pandas==1.2.4
streamlit==1.5.0
Pillow==9.0.0
scikit_learn==1.0.2
